+++ src/actions/video-comment-action.js
@@ -1,0 +1,6 @@
+export const updateComment = (comment) => {
+    return {
+        type: 'ADDED_COMMENT',
+        payload: comment
+    }
+};
