+++ src/actions/video-search-action.js
@@ -1,0 +1,16 @@
+import axios from 'axios';
+const YOUTUBE_URL = 'https://www.googleapis.com/youtube/v3/search';
+const API_KEY =  'AIzaSyDOkIHRmlpR0p2BCEvwePv8xlDP3YMo_UI';
+export const videoSearch = (term) => {
+    const params = {
+        part: 'snippet',
+        key: API_KEY,
+        term: term,
+        type: 'video',
+    };
+    const request = axios.get(YOUTUBE_URL, { params: params });
+    return {
+        type: "fetch",
+        payload: request
+      };
+}
