+++ src/reducers/index.js
@@ -1,11 +1,9 @@
 import {combineReducers} from 'redux';
 import InputFormReducer from './input-form-reducer';
 import SearchVideos from './search-videos';
-// import SelectedVideos from './selected-videos';
 
 const allReducers = combineReducers({
     comment: InputFormReducer,
-    term: SearchVideos,
-    // video: SelectedVideos
+    term: SearchVideos
 });
 export default allReducers;
