+++ src/reducers/search-videos.js
@@ -1,9 +1,10 @@
-import { SEARCH_VIDEOS } from './../actions/index';
 export default function(state = [], action) {
     switch (action.type) {
-      case 'SEARCH_VIDEOS':
-        return [...action.payload.items];
+      case 'fetch':
+        return action.payload.data.items; 
+        break;     
+        // return [...action.payload.items];
+      default: 
+        return state;
     }
-  
-    return state;
   }
